feat: restructure form pages to integrate submission history within form interface
- Create /a/[formId]/new page for form submission
- Convert /a/[formId] to smart redirect (latest submission or /new)
- Update /a/[formId]/submissions to handle both admin and user views
- Update /a/[formId]/submissions/[submissionId] for both admin and user views
- Remove My Submissions nav item from sidebar
- Redirect /my-submissions/* to new URL structure
